add code to support tree topology for scatter and test cases
