+++ lang/cs/Org.Apache.REEF.Tests/Functional/MPI/ScatterReduceTest/MasterTask.cs
@@ -50,8 +50,8 @@ namespace Org.Apache.REEF.Tests.Functional.MPI.ScatterReduceTest
         public byte[] Call(byte[] memento)
         {
             List<int> data = Enumerable.Range(1, 100).ToList();
-            List<string> order = GetScatterOrder();
-            _scatterSender.Send(data, order);
+            //List<string> order = GetScatterOrder();
+            _scatterSender.Send(data);
 
             int sum = _sumReducer.Reduce();
             _logger.Log(Level.Info, "Received sum: {0}", sum);
