+++ lang/cs/Org.Apache.REEF.Tests/Functional/MPI/ScatterReduceTest/ScatterReduceDriver.cs
@@ -30,6 +30,7 @@ using Org.Apache.REEF.Network.Group.Driver;
 using Org.Apache.REEF.Network.Group.Driver.Impl;
 using Org.Apache.REEF.Network.Group.Operators;
 using Org.Apache.REEF.Network.Group.Operators.Impl;
+using Org.Apache.REEF.Network.Group.Topology;
 using Org.Apache.REEF.Network.NetworkService;
 using Org.Apache.REEF.Tang.Annotations;
 using Org.Apache.REEF.Tang.Formats;
@@ -72,7 +73,8 @@ namespace Org.Apache.REEF.Tests.Functional.MPI.ScatterReduceTest
                         MpiTestConstants.ScatterOperatorName,
                         new ScatterOperatorSpec<int>(
                             MpiTestConstants.MasterTaskId,
-                            new IntCodec()))
+                            new IntCodec()),
+                            TopologyTypes.Tree)
                     .AddReduce(
                         MpiTestConstants.ReduceOperatorName,
                         new ReduceOperatorSpec<int>(
